requirements.txt
chromadb==1.3.4
gradio==5.49.1
llama-index==0.14.7
llama-index-embeddings-ollama==0.8.4
llama-index-llms-ollama==0.9.0
llama-index-vector-stores-chroma==0.5.3
ollama==0.6.0
openai-whisper==20250625
python-dotenv==1.2.1